+++ analyzer/config/hcl/testdata/deprecated.hcl
